+++ project/tests/test_management_commands.py
@@ -13,13 +13,6 @@ def test_envhelp_works():
     assert 'DEBUG' in out.getvalue()
 
 
-def test_sendtestsms_works(smsoutbox):
-    call_command('sendtestsms', '5551234568', 'blarg')
-    assert len(smsoutbox) == 1
-    assert smsoutbox[0].to == '+15551234568'
-    assert smsoutbox[0].body == 'blarg'
-
-
 class SendtestslackTests(TestCase):
     @override_settings(SLACK_WEBHOOK_URL='')
     def test_it_raises_error_when_settings_are_not_defined(self):
