+++ texting/apps.py
@@ -3,3 +3,8 @@ from django.apps import AppConfig
 
 class TextingConfig(AppConfig):
     name = 'texting'
+
+    def ready(self):
+        from . import twilio
+
+        twilio.validate_settings()
