+++ texting/tests/test_sendtestsms.py
@@ -1,0 +1,8 @@
+from django.core.management import call_command
+
+
+def test_sendtestsms_works(smsoutbox):
+    call_command('sendtestsms', '5551234568', 'blarg')
+    assert len(smsoutbox) == 1
+    assert smsoutbox[0].to == '+15551234568'
+    assert smsoutbox[0].body == 'blarg'
